+++ relay-dynamic-config/src/feature.rs
@@ -61,11 +61,6 @@ pub enum Feature {
     /// Serialized as `organizations:relay-otel-logs-endpoint`.
     #[serde(rename = "organizations:relay-otel-logs-endpoint")]
     OtelLogsEndpoint,
-    /// Enable logs ingestion via the Vercel Log Drain endpoint.
-    ///
-    /// Serialized as `organizations:relay-vercel-log-drain-endpoint`.
-    #[serde(rename = "organizations:relay-vercel-log-drain-endpoint")]
-    VercelLogDrainEndpoint,
     /// Enable playstation crash dump ingestion via the `/playstation/` endpoint.
     ///
     /// Serialized as `organizations:relay-playstation-ingestion`.